requirements.txt
fastapi
uvicorn
pydantic
firebase-admin
python-dotenv
google-generativeai
yt-dlp